debug delete & API KEY Activation

+++ api/delete.js
@@ -1,4 +1,4 @@
-const S3DB = require('@dashpilot/s3-json-db');
+const S3DB = require('../s3-json-db');
 const db = new S3DB(process.env.ACCESS_KEY, process.env.SECRET_KEY, 'myfiles', null, null, 'gateway.ap1.storjshare.io');
 const table = 'entries';
 module.exports = async function(req,res){
@@ -7,6 +7,11 @@ module.exports = async function(req,res){
         res.status(400).send("id param missing in POST body")
         return;
     }
+    if(req.query.API_KEY!=process.env.API_KEY){
+        res.status(401).send("Invalid API Key")
+        return;
+    }
+
    db.delete(table, req.body.id).then(id => {
     res.send("successfully deleted ",id)
